Login Version 2.2

+++ PCUMS/Login1.cs
@@ -21,36 +21,51 @@ namespace PCUMS
     {
 
         public bool UserCredentialsCreated { get; private set; }
-        public Login1(bool Value)
+        public Login1()
         {
             Program.Requester = 0;
             InitializeComponent();
-            if (Value == true)
-            {
-                userName.Enabled = false;
-                passWord.Enabled = false;
-                save1.Enabled = false;
-            }
+           
             
         }
 
         private void save1_Click(object sender, EventArgs e)
         {
-                if (String.IsNullOrEmpty(userName.Text) == false && String.IsNullOrEmpty(passWord.Text) == false
+           if (String.IsNullOrEmpty(userName.Text) == false && String.IsNullOrEmpty(passWord.Text) == false
                 && passWord.Text.Contains(" ") == false && userName.Text.Contains(" ") == false)
-                {
-                    
+           {
+               if (!File.Exists(Program.credentialsPath))
+               {
+                    Program.AdminID = "1";
                     Program.csv = Program.AdminID + "," + userName.Text + "," + passWord.Text;
                     System.IO.File.WriteAllText(Program.credentialsPath, Program.csv);
                     userName.Enabled = false;
                     passWord.Enabled = false;
                     save1.Enabled = false;
+               }
+               else
+               {
+                    String temp = File.ReadLines(Program.credentialsPath).Last();
+                    int tempInt = 0;
+                    temp = temp.Split(',')[0];
+                    tempInt = Int32.Parse(temp);
+                    tempInt = tempInt + 1;
+                    temp = tempInt.ToString();
+                    Program.AdminID = temp;
+                    Program.csv = Program.AdminID + "," + userName.Text + "," + passWord.Text;
+                    File.AppendAllText(Program.credentialsPath, Environment.NewLine + Program.csv);
+                    userName.Enabled = false;
+                    passWord.Enabled = false;
+                    save1.Enabled = false;
+               }
+
+                
                     
-                }
-                else
-                {
-                    System.Windows.Forms.MessageBox.Show("Create admin credentials and make sure there are no whitespaces!");
-                }
+           }
+           else
+           {
+           System.Windows.Forms.MessageBox.Show("Create admin credentials and make sure there are no whitespaces!");
+           }
             
         }
 
@@ -89,9 +104,10 @@ namespace PCUMS
                 string store = System.IO.File.ReadAllText(Program.credentialsPath);
                 string admin = store.Split(',')[1];
             
-                userName.Enabled = false;
-                passWord.Enabled = false;
-                save1.Enabled = false;
+               // userName.Enabled = false;
+              //  passWord.Enabled = false;
+              //  save1.Enabled = false;
+
                 
                 label2.Text = "Hello " + admin + ", let us get started";
             }
